Move E2E test to existing mpa_v2_basics test file
Integrate the Unicode URL path test into the existing MPA v2 basics test
file rather than creating separate test files.
