--- e2e_playwright/multipage_apps_v2/mpa_v2_unicode_paths.py
@@ -1,41 +1,0 @@
-# Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2026)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Test app for Unicode URL path navigation (regression test for GH issue 15267)."""
-
-import streamlit as st
-
-
-def home_page():
-    st.header("Home")
-    st.write("This is the home page.")
-
-
-def unicode_page():
-    st.header("Págé_Wíth_Spêcîãl_Chäracters")
-    st.write("This page has Unicode characters in the URL path.")
-
-
-pg = st.navigation(
-    [
-        st.Page(home_page, title="Home", url_path="home", default=True),
-        st.Page(
-            unicode_page,
-            title="Págé_Wíth_Spêcîãl_Chäracters",
-            url_path="Págé_Wíth_Spêcîãl_Chäracters",
-        ),
-    ]
-)
-
-pg.run()
